Refactor TabularVPGActorCritic into separate module
Also create a base Tabular class that can be also be used by HCA Actor Critics
